+++ lib/experimental/Mockups/menu1.ts
@@ -1,0 +1,129 @@
+import * as Icons from "@/icons/app"
+import { rootItems } from "./menu"
+
+// Full CrossFit
+
+const personalItems = [
+  { label: "Profile", icon: Icons.Person, href: "/profile" },
+  {
+    label: "Tasks",
+    icon: Icons.CheckCircleLine,
+    href: "/tasks",
+  },
+  {
+    label: "My documents",
+    icon: Icons.Folder,
+    href: "/my-documents",
+  },
+]
+
+export const menuTreeNone = [
+  {
+    title: "Root",
+    id: "root",
+    items: rootItems,
+    isRoot: true,
+  },
+  {
+    title: "Personal",
+    id: "personal",
+    items: personalItems,
+    isOpen: true,
+  },
+  {
+    title: "Company",
+    id: "company",
+    items: [
+      {
+        label: "Organization",
+        icon: Icons.People,
+        href: "/organization",
+      },
+      {
+        label: "Calendar",
+        icon: Icons.Calendar,
+        href: "/calendar",
+      },
+      {
+        label: "Documents",
+        icon: Icons.Folder,
+        href: "/documents",
+      },
+      {
+        label: "Analytics",
+        icon: Icons.BarGraph,
+        href: "/analytics",
+      },
+      {
+        label: "Workflows",
+        icon: Icons.Split,
+        href: "/workflows",
+      },
+      {
+        label: "Settings",
+        icon: Icons.Settings,
+        href: "/settings",
+      },
+    ],
+    isOpen: true,
+  },
+]
+
+export const menuTreeBundles = [
+  {
+    title: "Root",
+    id: "root",
+    items: rootItems,
+    isRoot: true,
+  },
+  {
+    title: "Personal",
+    id: "personal",
+    items: personalItems,
+    isOpen: true,
+  },
+  {
+    title: "General",
+    id: "general",
+    items: [
+      {
+        label: "Organization",
+        icon: Icons.People,
+        href: "/organization",
+      },
+      {
+        label: "Calendar",
+        icon: Icons.Calendar,
+        href: "/calendar",
+      },
+      {
+        label: "Documents",
+        icon: Icons.Folder,
+        href: "/documents",
+      },
+    ],
+    isOpen: true,
+  },
+  {
+    title: "More",
+    id: "more",
+    items: [
+      {
+        label: "Analytics",
+        icon: Icons.BarGraph,
+        href: "/analytics",
+      },
+      {
+        label: "Workflows",
+        icon: Icons.Split,
+        href: "/workflows",
+      },
+      {
+        label: "Settings",
+        icon: Icons.Settings,
+        href: "/settings",
+      },
+    ],
+    isOpen: true,
+  },
+]
